feat(langflow): add function support to build_template_from_class function feat(langflow): add function base class to InitializeAgentNode
The `build_template_from_class` function in `ChainCreator` now supports adding a function to the chain. This is done by passing `add_function=True` to the function.

In `InitializeAgentNode`, the `base_classes` attribute has been updated to include the `function` base class. This allows the node to be used as a function in the chain.