feat: Add dynamic field to code field in types module
This commit adds a dynamic field to the code field in the types module. The dynamic field is set to True, indicating that the field allows updates.

refactor: Remove dynamic field from FrontendNode class

In the FrontendNode class of the base module, the dynamic field has been removed. This change aligns with the removal of the dynamic field in the types module.

refactor: Remove dynamic field from CustomComponentNode class

The dynamic field has been removed from the CustomComponentNode class in the tools module. This aligns with the removal of the dynamic field in the FrontendNode class.